Optimizing for LWRP
Big changes include switching trees to "Baked Lit" shader instead of SpeedTrees8 or whatever; that shader wasn't compatible with SRP, so we were losing out on some batching. With new occlusion strategies everything usually gets above 50 fps (with 2 dynamic/mixed light sources , no shadows, and MSAA x2), but performance drops a ton when looking into the sky; probably because many many trees are being drawn, each one that is part of the viewable canopy. I think a great fix might be seperating the trunks from the foliage and placing foliage manually - the idea would be to try and capitilze on the "instancing" strategy (maybe not "true" instancing as understood in other batcher) while also capitilzing on frustrum culling. Still need to experiment with 2019.3; if single-pass forward rendering works with the Quest (even at the cost of being able to use the terrain editor) I think it might put the framereate over the 72 per second mark.